DRY up the Serializers
